always use int64 for counters

+++ internal/handler/mem_service.go
@@ -4,6 +4,6 @@ type MemStorage interface {
 	Gauges() map[string]float64
 	Counters() map[string]int64
 	UpdateGauge(name string, value float64) error
-	UpdateCounter(name string, value int) error
+	UpdateCounter(name string, value int64) error
 	OverrideCounter(name string, value int64) error
 }
